+++ demo/App.vue
@@ -1,62 +1,75 @@
 <template>
   <div>
-    <base-header />
+    <base-header v-bind="vsmProps" />
     <main>
-      <install-docs />
-      <example-docs />
-      <example-codepen-docs />
-      <props-menu-docs />
-      <menu-object-menu-docs />
-      <slots-menu-docs />
-      <events-menu-docs />
-      <methods-menu-docs />
-      <properties-menu-docs />
-      <props-mob-docs />
-      <slots-mob-docs />
-      <methods-mob-docs />
-      <classes-docs />
-      <demo-docs />
+      <conf-menu-docs
+        @on-css="onChangeCss"
+        @change-props="onChangeProps"
+      />
+      <install-docs
+        :css="css"
+        :vsm-props="vsmProps"
+      />
+      <full-example-docs />
+      <properties-docs
+        v-for="property in properties"
+        :key="property.title"
+        v-bind="property"
+      />
     </main>
     <base-footer />
   </div>
 </template>
 
 <script>
-import MenuObjectMenuDocs from './components/docs/menu/MenuObject'
-import PropertiesMenuDocs from './components/docs/menu/Properties'
-import ExampleCodepenDocs from './components/docs/ExampleCodepen'
-import MethodsMenuDocs from './components/docs/menu/Methods'
-import MethodsMobDocs from './components/docs/mob/Methods'
-import EventsMenuDocs from './components/docs/menu/Events'
-import PropsMenuDocs from './components/docs/menu/Props'
-import SlotsMenuDocs from './components/docs/menu/Slots'
-import PropsMobDocs from './components/docs/mob/Props'
-import SlotsMobDocs from './components/docs/mob/Slots'
-import ClassesDocs from './components/docs/Classes'
-import ExampleDocs from './components/docs/Example'
+import FullExampleDocs from './components/docs/FullExample'
+import PropertiesDocs from './components/docs/Properties'
+import ConfMenuDocs from './components/docs/ConfMenu'
 import InstallDocs from './components/docs/Install'
 import BaseHeader from './components/BaseHeader'
 import BaseFooter from './components/BaseFooter'
-import DemoDocs from './components/docs/Demo'
+
+import * as menuEvents from './docs/menu/events'
+import * as menuObject from './docs/menu/menuObject'
+import * as menuMethods from './docs/menu/methods'
+import * as menuProperties from './docs/menu/properties'
+import * as menuProps from './docs/menu/props'
+import * as menuSlots from './docs/menu/slots'
+import * as mobMethods from './docs/mob/methods'
+import * as mobProps from './docs/mob/props'
+import * as mobSlots from './docs/mob/slots'
+import * as classes from './docs/classes'
 
 export default {
   components: {
-    MenuObjectMenuDocs,
-    PropertiesMenuDocs,
-    ExampleCodepenDocs,
-    MethodsMenuDocs,
-    MethodsMobDocs,
-    EventsMenuDocs,
-    PropsMenuDocs,
-    SlotsMenuDocs,
-    PropsMobDocs,
-    SlotsMobDocs,
-    ClassesDocs,
-    ExampleDocs,
+    FullExampleDocs,
+    PropertiesDocs,
+    ConfMenuDocs,
     InstallDocs,
     BaseHeader,
-    BaseFooter,
-    DemoDocs
+    BaseFooter
+  },
+  setup() {
+    const properties = [
+      menuProps, menuObject, menuEvents, menuSlots, menuMethods, menuProperties,
+      mobProps, mobSlots, mobMethods, classes
+    ]
+
+    return { properties }
+  },
+  data() {
+    return {
+      css: '',
+      vsmProps: {}
+    }
+  },
+  methods: {
+    onChangeProps(obj) {
+      this.vsmProps = obj
+    },
+    onChangeCss(val) {
+      this.css = val
+    }
   }
 }
 </script>