+++ src/components/Menu.vue
@@ -3,17 +3,19 @@
     :is="element"
     class="vsm-menu vsm-no-transition"
   >
-    <nav>
+    <nav class="vsm-nav">
       <ul
         ref="root"
         class="vsm-root"
       >
         <slot name="before-nav" />
-        <li class="vsm-section vsm-section_menu vsm-mob-hide">
+        <li
+          ref="linkContainer"
+          class="vsm-link-container vsm-mob-hide"
+        >
           <component
             :is="item.element || (item.dropdown ? 'button' : 'a')"
             v-for="(item, index) in menu"
-            ref="links"
             :key="index"
             :class="['vsm-link', item.attributes ? item.attributes.class : null, {
               'vsm-has-dropdown': item.dropdown
@@ -21,8 +23,9 @@
             :data-dropdown="item.dropdown"
             :aria-haspopup="item.dropdown && 'true'"
             :aria-expanded="item.dropdown && 'false'"
+            tabindex="0"
             v-bind="item.attributes"
-            v-on="item.listeners"
+            v-on="item.listeners || {}"
           >
             <slot
               name="title"
@@ -55,8 +58,7 @@
         class="vsm-dropdown-container"
       >
         <div
-          v-for="(item, index) in menuHasDropdown"
-          ref="sections"
+          v-for="(item, index) in itemsWithDropdown"
           :key="index"
           class="vsm-dropdown-section"
           :data-dropdown="item.dropdown"
@@ -75,21 +77,21 @@
 </template>
 
 <script>
-import Vue from 'vue'
+// This values depends on .vsm-background styles (width/height)
+const BASE_WIDTH = 380;
+const BASE_HEIGHT = 400;
 
 export default {
   name: 'VsmMenu',
   props: {
     /**
-     * An array of objects that, when initialized,
-     * turn into HTML elements
      * @example
      *  [{
-     *   // display menu item
+     *   // display menu item, can be overridden with slots
      *   title: 'News',
      *   // activate dropdown content, must be unique!
      *   dropdown: 'news',
-     *   // change the HTML element to ours
+     *   // change the default HTMLElement (element/global component)
      *   element: 'router-link',
      *   // v-bind accepts
      *   attributes: {
@@ -101,7 +103,7 @@ export default {
      *     mouseover: (evt) => console.log('news hover', evt)
      *   },
      *   // other attributes
-     *   new_item: true,
+     *   customAttribute: true,
      *  }]
      * },
      */
@@ -110,28 +112,13 @@ export default {
       required: true
     },
     /**
-     * Change root HTML element
-     * @example
-     *  div
+     * Change root HTMLElement
+     * @example div, section
      */
     element: {
       type: String,
-      default: 'header'
-    },
-    /**
-     * Problems displaying the menu? Try changing the two lower
-     * properties to the average width and height
-     * of your dropdown content.
-     */
-    baseWidth: {
-      type: [Number, String],
-      default: 380,
-      validator: (val) => +val > 0
-    },
-    baseHeight: {
-      type: [Number, String],
-      default: 400,
-      validator: (val) => +val > 0
+      default: 'header',
+      validator: (val) => !!val
     },
     /**
      * Dropdown content does not go beyond screen size
@@ -142,6 +129,11 @@ export default {
       default: 10,
       validator: (val) => +val >= 0
     },
+    dropdownOffset: {
+      type: [Number, String],
+      default: 0,
+      validator: (val) => +val >= 0
+    },
     /**
      * By default, the dropdown list drops out on hover,
      * you can change this behavior on click
@@ -150,228 +142,139 @@ export default {
       type: String,
       default: 'hover',
       validator: (val) => ['hover', 'click'].includes(val)
+    },
+    /**
+     * Must be equals as $vsm-transition (scss variable)
+     * @default .25s
+     */
+    transitionTimeout: {
+      type: [Number, String],
+      default: 250
+    }
+  },
+  emits: [
+    'open-dropdown', 'close-dropdown'
+  ],
+  data() {
+    return {
+      elementsWithDropdown: [],
+      dropdownContainerItems: []
     }
   },
   computed: {
-    /**
-     * Menu items that have dropdown content
-     */
-    menuHasDropdown () {
+    itemsWithDropdown() {
       return this.menu.filter(item => item.dropdown)
-    },
-    /**
-     * HTML menu elements that have dropdown content
-     */
-    hasDropdownEls () {
-      const links = this.$refs.links || []
-      const elements = []
-
-      links.forEach((link) => {
-        const el = link instanceof Vue ? link.$el : link
-
-        if (el.classList.contains('vsm-has-dropdown')) {
-          elements.push(el)
-        }
-      })
-
-      return elements
-    },
-    /**
-     * HTML dropdown content
-     */
-    sectionEls () {
-      const sections = this.$refs.sections || []
-
-      return sections.map((el) => ({
-        el,
-        name: el.getAttribute('data-dropdown'),
-        content: el.children[0]
-      }))
     }
   },
   watch: {
-    handler (val) {
-      this.handler = val
-      this.registerDropdownElsEvents(true)
+    async element() {
+      await this.$nextTick()
+      this.updateDataElements()
+      this.registerDropdownElementsEvents()
+    },
+    handler() {
+      this.registerDropdownElementsEvents(true)
       this.registerDropdownContainerEvents(true)
+    },
+    menu: {
+      async handler() {
+        await this.$nextTick()
+        this.updateDataElements()
+        this.registerDropdownElementsEvents()
+      },
+      deep: true
     }
   },
-  mounted () {
-    // PointerEvent interface represents the state of a DOM event
-    this._pointerEvent = window.PointerEvent ? {
-      end: 'pointerup',
-      enter: 'pointerenter',
-      leave: 'pointerleave'
-    } : {
-      end: 'touchend',
-      enter: 'mouseenter',
-      leave: 'mouseleave'
-    }
-
-    this.registerGlobalEvents()
-    this.registerDropdownElsEvents()
+  mounted() {
+    this.identifyPointerEvents()
+    this.registerGlobalListeners()
+    this.updateDataElements()
+    this.registerDropdownElementsEvents()
     this.registerDropdownContainerEvents()
   },
-  beforeDestroy () {
-    this.unregisterGlobalEvents()
+  beforeUnmount() {
+    this.removeGlobalListeners()
   },
   methods: {
-    registerGlobalEvents () {
-      window.addEventListener('resize', this.windowResizeHandler)
-      document.addEventListener('touchmove', this.touchMoveHandler)
-      document.addEventListener('touchstart', this.touchStartHandler)
-      document.body.addEventListener(this._pointerEvent.end, this.eventEndHandler)
-    },
-    registerDropdownElsEvents (force = false) {
-      this.hasDropdownEls.forEach((el) => {
-        // Events have been registered
-        if (el._vsmMenu && !force) {
-          return
-        }
-
-        if (el._vsmMenuHandlers) {
-          Object.entries(el._vsmMenuHandlers).forEach(([eventName, fn]) => {
-            el.removeEventListener(eventName, fn)
-          })
-        }
-
-        if (this.handler === 'hover') {
-          el._vsmMenuHandlers = {
-            focusin: () => {
-              this.stopCloseTimeout()
-              this.openDropdown(el)
-            },
-            [this._pointerEvent.enter]: (evt) => {
-              if (evt.pointerType !== 'touch') {
-                this.stopCloseTimeout()
-                this.openDropdown(el)
-              }
-            },
-            [this._pointerEvent.leave]: (evt) => {
-              if (evt.pointerType !== 'touch') {
-                this.startCloseTimeout()
-              }
-            }
-          }
-        } else {
-          el._vsmMenuHandlers = {}
-        }
-
-        el._vsmMenuHandlers[this._pointerEvent.end] = (evt) => {
-          evt.preventDefault()
-          evt.stopPropagation()
-          this.toggleDropdown(el)
-        }
-
-        Object.entries(el._vsmMenuHandlers).forEach(([eventName, fn]) => {
-          el.addEventListener(eventName, fn)
-        })
-
-        el._vsmMenu = true
-      })
-    },
-    registerDropdownContainerEvents (force = false) {
-      const el = this.$refs.dropdownContainer
-
-      // Events have been registered
-      if (el._vsmMenu && !force) {
-        return
-      }
-
-      if (el._vsmMenuHandlers) {
-        Object.entries(el._vsmMenuHandlers).forEach(([eventName, fn]) => {
-          el.removeEventListener(eventName, fn)
-        })
-      }
-
-      if (this.handler === 'hover') {
-        el._vsmMenuHandlers = {
-          [this._pointerEvent.enter]: (evt) => {
-            if (evt.pointerType !== 'touch') {
-              this.stopCloseTimeout()
-            }
-          },
-          [this._pointerEvent.leave]: (evt) => {
-            if (evt.pointerType !== 'touch') {
-              this.startCloseTimeout()
-            }
-          }
-        }
-      } else {
-        el._vsmMenuHandlers = {}
-      }
-
-      el._vsmMenuHandlers[this._pointerEvent.end] = (evt) => {
-        evt.stopPropagation()
-      }
-
-      Object.entries(el._vsmMenuHandlers).forEach(([eventName, fn]) => {
-        el.addEventListener(eventName, fn)
-      })
-
-      el._vsmMenu = true
-    },
-    unregisterGlobalEvents () {
-      window.removeEventListener('resize', this.windowResizeHandler)
-      document.removeEventListener('touchmove', this.touchMoveHandler)
-      document.removeEventListener('touchstart', this.touchStartHandler)
-      document.body.removeEventListener(this._pointerEvent.end, this.eventEndHandler)
-    },
-    toggleDropdown (el) {
+    /*
+     * | ------------------------------------------------------------------------------------------------
+     * | - Main Functions -
+     * | ------------------------------------------------------------------------------------------------
+     */
+    toggleDropdown(el) {
       if (this._activeDropdown === el) {
         this.closeDropdown()
       } else {
         this.openDropdown(el)
       }
     },
-    openDropdown (el) {
+    openDropdown(el) {
       if (this._activeDropdown === el) {
         return
       }
 
       this.$emit('open-dropdown', el)
 
-      this.$el.classList.add('vsm-overlay-active')
-      this.$el.classList.add('vsm-dropdown-active')
+      this.$el.classList.add('vsm-overlay-active', 'vsm-dropdown-active')
       this._activeDropdown = el
       this._activeDropdown.setAttribute('aria-expanded', 'true')
-      this.hasDropdownEls.forEach(el => el.classList.remove('vsm-active'))
+      this.elementsWithDropdown.forEach(el => el.classList.remove('vsm-active'))
       el.classList.add('vsm-active')
 
       const activeDataDropdown = el.getAttribute('data-dropdown')
       let direction = 'vsm-left'
-      let offsetWidth
-      let offsetHeight
-      let content
 
-      this.sectionEls.forEach((item) => {
-        item.el.classList.remove('vsm-active')
-        item.el.classList.remove('vsm-left')
-        item.el.classList.remove('vsm-right')
+      this.dropdownContainerItems.forEach((item) => {
+        item.el.classList.remove('vsm-active', 'vsm-left', 'vsm-right')
 
         if (item.name === activeDataDropdown) {
           item.el.setAttribute('aria-hidden', 'false')
           item.el.classList.add('vsm-active')
           direction = 'vsm-right'
-          offsetWidth = item.content.offsetWidth
-          offsetHeight = item.content.offsetHeight
-          content = item.content
+          this._activeContainerItem = item
         } else {
-          item.el.classList.add(direction)
           item.el.setAttribute('aria-hidden', 'true')
+          item.el.classList.add(direction)
         }
       })
 
+      this.resizeDropdown()
+    },
+    closeDropdown() {
+      if (!this._activeDropdown) {
+        return
+      }
+
+      this.$emit('close-dropdown', this._activeDropdown)
+      this.elementsWithDropdown.forEach((el) => el.classList.remove('vsm-active'))
+
+      this._activeContainerItem.el.setAttribute('aria-hidden', 'true')
+
+      this.clearEnableTransitionTimeout()
+      this.startDisableTransitionTimeout()
+
+      this.$el.classList.remove('vsm-overlay-active', 'vsm-dropdown-active')
+
+      this._activeDropdown.setAttribute('aria-expanded', 'false')
+      this._activeContainerItem = undefined
+      this._activeDropdown = undefined
+    },
+    resizeDropdown() {
+      if (!this._activeContainerItem) {
+        return
+      }
+
       const bodyWidth = document.documentElement.offsetWidth
       const rootRect = this.$el.getBoundingClientRect()
-      const rect = el.getBoundingClientRect()
+      const rect = this._activeDropdown.getBoundingClientRect()
+
+      let { offsetHeight, offsetWidth } = this._activeContainerItem.content
 
       // Find the beginning of a menu item
       const leftPosition = rect.left - rootRect.left
 
-      // Step back from the button to the left so that
-      // the middle of the content is found in the
-      // center of the element
+      // Step back from the button to the left so that the middle of
+      // the content is found in the center of the element
       let centerPosition = leftPosition - (offsetWidth / 2) + (rect.width / 2)
 
       // Do not let go of the left side of the screen
@@ -395,90 +298,224 @@ export default {
       // Possible blurring font with decimal values
       centerPosition = Math.round(centerPosition)
 
-      const ratioWidth = offsetWidth / +this.baseWidth
-      const ratioHeight = offsetHeight / +this.baseHeight
+      const dropdownOffset = +this.dropdownOffset + this._activeDropdown.offsetTop
+      const ratioWidth = offsetWidth / BASE_WIDTH
+      const ratioHeight = offsetHeight / BASE_HEIGHT
 
       // Activate transition
-      clearTimeout(this._disableTransitionTimeout)
-      this._enableTransitionTimeout = setTimeout(() => {
-        this.$el.classList.remove('vsm-no-transition')
-      }, 50)
+      this.clearDisableTransitionTimeout()
+      this.startEnableTransitionTimeout()
 
-      this.$refs.dropdownContainer.style.transform = `translate(${centerPosition}px, ${el.offsetTop}px)`
+      this.$refs.dropdownContainer.style.transform = `translate(${centerPosition}px, ${dropdownOffset}px)`
       this.$refs.dropdownContainer.style.width = `${offsetWidth}px`
       this.$refs.dropdownContainer.style.height = `${offsetHeight}px`
 
-      this.$refs.arrow.style.transform = `translate(${leftPosition + (rect.width / 2)}px, ${el.offsetTop}px) rotate(45deg)`
-      this.$refs.background.style.transform = `translate(${centerPosition}px, ${el.offsetTop}px) scaleX(${ratioWidth}) scaleY(${ratioHeight})`
-      this.$refs.backgroundAlt.style.transform = `translateY(${content.children[0].offsetHeight / ratioHeight}px)`
+      this.$refs.arrow.style.transform = `translate(${leftPosition + (rect.width / 2)}px, ${dropdownOffset}px) rotate(45deg)`
+      this.$refs.background.style.transform = `translate(${centerPosition}px, ${dropdownOffset}px) scaleX(${ratioWidth}) scaleY(${ratioHeight})`
+      this.$refs.backgroundAlt.style.transform = `translateY(${this._activeContainerItem.content.firstElementChild.offsetHeight / ratioHeight}px)`
     },
-    closeDropdown () {
-      if (!this._activeDropdown) {
+    /*
+     * | ------------------------------------------------------------------------------------------------
+     * | - Timeout -
+     * | ------------------------------------------------------------------------------------------------
+     */
+    startCloseDropdownTimeout() {
+      this._closeDropdownTimeout = setTimeout(() => this.closeDropdown(), 50)
+    },
+    clearCloseDropdownTimeout() {
+      clearTimeout(this._closeDropdownTimeout)
+    },
+    startEnableTransitionTimeout() {
+      this._enableTransitionTimeout = setTimeout(() => this.$el.classList.remove('vsm-no-transition'), 50)
+    },
+    clearEnableTransitionTimeout() {
+      clearTimeout(this._enableTransitionTimeout)
+    },
+    startDisableTransitionTimeout() {
+      this._disableTransitionTimeout = setTimeout(() => this.$el.classList.add('vsm-no-transition'), +this.transitionTimeout)
+    },
+    clearDisableTransitionTimeout() {
+      clearTimeout(this._disableTransitionTimeout)
+    },
+    /*
+     * | ------------------------------------------------------------------------------------------------
+     * | - Events -
+     * | ------------------------------------------------------------------------------------------------
+     */
+    registerDropdownElementsEvents(force = false) {
+      this.elementsWithDropdown.forEach((el) => {
+        // Events have been registered
+        if (el._vsmMenu && !force) {
+          return
+        }
+
+        if (el._vsmMenuHandlers) {
+          Object.entries(el._vsmMenuHandlers).forEach(([eventName, fn]) => {
+            el.removeEventListener(eventName, fn)
+          })
+        }
+
+        if (this.handler === 'hover') {
+          el._vsmMenuHandlers = {
+            focusin: () => {
+              this.clearCloseDropdownTimeout()
+              this.openDropdown(el)
+            },
+            [this._pointerEvent.enter]: (evt) => {
+              if (evt.pointerType !== 'touch') {
+                this.clearCloseDropdownTimeout()
+                this.openDropdown(el)
+              }
+            },
+            [this._pointerEvent.leave]: (evt) => {
+              if (evt.pointerType !== 'touch') {
+                this.startCloseDropdownTimeout()
+              }
+            }
+          }
+        } else {
+          el._vsmMenuHandlers = {}
+        }
+
+        el._vsmMenuHandlers[this._pointerEvent.end] = (evt) => {
+          evt.preventDefault()
+          evt.stopPropagation()
+          this.toggleDropdown(el)
+        }
+
+        Object.entries(el._vsmMenuHandlers).forEach(([eventName, fn]) => {
+          el.addEventListener(eventName, fn)
+        })
+
+        el._vsmMenu = true
+      })
+    },
+    registerDropdownContainerEvents(force = false) {
+      const el = this.$refs.dropdownContainer
+
+      // Events have been registered
+      if (el._vsmMenu && !force) {
         return
       }
 
-      this.$emit('close-dropdown', this._activeDropdown)
+      if (el._vsmMenuHandlers) {
+        Object.entries(el._vsmMenuHandlers).forEach(([eventName, fn]) => {
+          el.removeEventListener(eventName, fn)
+        })
+      }
 
-      this.hasDropdownEls.forEach((el) => {
-        el.classList.remove('vsm-active')
+      if (this.handler === 'hover') {
+        el._vsmMenuHandlers = {
+          [this._pointerEvent.enter]: (evt) => {
+            if (evt.pointerType !== 'touch') {
+              this.clearCloseDropdownTimeout()
+            }
+          },
+          [this._pointerEvent.leave]: (evt) => {
+            if (evt.pointerType !== 'touch') {
+              this.startCloseDropdownTimeout()
+            }
+          }
+        }
+      } else {
+        el._vsmMenuHandlers = {}
+      }
+
+      el._vsmMenuHandlers[this._pointerEvent.end] = (evt) => {
+        evt.stopPropagation()
+      }
+
+      Object.entries(el._vsmMenuHandlers).forEach(([eventName, fn]) => {
+        el.addEventListener(eventName, fn)
       })
 
-      const activeDropdownSection = this.$refs.dropdownContainer.querySelector('[aria-hidden="false"]')
-      if (activeDropdownSection) {
-        activeDropdownSection.setAttribute('aria-hidden', 'true')
-      }
-
-      clearTimeout(this._enableTransitionTimeout)
-
-      this._disableTransitionTimeout = setTimeout(() => {
-        this.$el.classList.add('vsm-no-transition')
-      }, 50)
-
-      this.$el.classList.remove('vsm-overlay-active')
-      this.$el.classList.remove('vsm-dropdown-active')
-
-      this._activeDropdown.setAttribute('aria-expanded', 'false')
-      this._activeDropdown = undefined
-    },
-    startCloseTimeout () {
-      this._closeDropdownTimeout = setTimeout(() => {
-        this.closeDropdown()
-      }, 50)
-    },
-    stopCloseTimeout () {
-      clearTimeout(this._closeDropdownTimeout)
-    },
-    touchMoveHandler () {
-      this._isDragging = true
-    },
-    touchStartHandler () {
-      this._isDragging = false
-    },
-    eventEndHandler () {
-      if (!this._isDragging) {
-        this.closeDropdown()
-      }
+      el._vsmMenu = true
     },
     /*
-     * When the screen is reduced, the active drop-down content
-     * does not change its size, because of this,
-     * horizontal scrolling may occur
+     * | ------------------------------------------------------------------------------------------------
+     * | - Global Listener Handlers -
+     * | ------------------------------------------------------------------------------------------------
      */
-    windowResizeHandler () {
-      // Block dropdown from closing on scroll
-      // (show/hide browser top header)
+    registerGlobalListeners() {
+      window.addEventListener('resize', this.windowResizeHandler)
+      document.addEventListener('touchmove', this.documentTouchMoveHandler)
+      document.addEventListener('touchstart', this.documentTouchStartHandler)
+      document.body.addEventListener(this._pointerEvent.end, this.documentEventEndHandler)
+    },
+    removeGlobalListeners() {
+      window.removeEventListener('resize', this.windowResizeHandler)
+      document.removeEventListener('touchmove', this.documentTouchMoveHandler)
+      document.removeEventListener('touchstart', this.documentTouchStartHandler)
+      document.body.removeEventListener(this._pointerEvent.end, this.documentEventEndHandler)
+    },
+    windowResizeHandler() {
+      // Recalculates the dropdown only in cases where the screen width changes
       if (this._lastWindowWidth === window.innerWidth) {
         return
       }
 
       this._lastWindowWidth = window.innerWidth
 
-      this.$refs.dropdownContainer.style = null
-      this.$refs.arrow.style = null
-      this.$refs.background.style = null
-      this.$refs.backgroundAlt.style = null
+      if (this._activeDropdown) {
+        this.$el.classList.add('vsm-no-transition')
+        this.resizeDropdown()
+        return
+      }
 
+      // Don't do unnecessary actions
+      if (!this.$refs.background.getAttribute('style')) {
+        return
+      }
+
+      // vsm-no-transition does not apply to dropdown (which has an opacity property),
+      // with a long animation you can see unnecessary animation
+      this.$el.classList.add('vsm-no-transition', 'vsm-no-transition_dropdown')
+      setTimeout(() => this.$el.classList.remove('vsm-no-transition_dropdown'))
+
+      // The dropdown position may remain on the right side of the screen,
+      // causing a horizontal scroll
       this.closeDropdown()
+      this.clearAllStyles()
+    },
+    documentTouchMoveHandler() {
+      this._isDragging = true
+    },
+    documentTouchStartHandler() {
+      this._isDragging = false
+    },
+    documentEventEndHandler() {
+      if (!this._isDragging) {
+        this.closeDropdown()
+      }
+    },
+    /*
+     * | ------------------------------------------------------------------------------------------------
+     * | - Utils -
+     * | ------------------------------------------------------------------------------------------------
+     */
+    identifyPointerEvents() {
+      this._pointerEvent = window.PointerEvent ? {
+        end: 'pointerup',
+        enter: 'pointerenter',
+        leave: 'pointerleave'
+      } : {
+        end: 'touchend',
+        enter: 'mouseenter',
+        leave: 'mouseleave'
+      }
+    },
+    updateDataElements() {
+      this.elementsWithDropdown = Array.from(this.$refs.linkContainer.children)
+        .filter((el) => el.classList.contains('vsm-has-dropdown'))
+
+      this.dropdownContainerItems = Array.from(this.$refs.dropdownContainer.children)
+        .map((el) => ({ el, name: el.getAttribute('data-dropdown'), content: el.firstElementChild }))
+    },
+    clearAllStyles() {
+      this.$refs.dropdownContainer.removeAttribute('style')
+      this.$refs.arrow.removeAttribute('style')
+      this.$refs.background.removeAttribute('style')
+      this.$refs.backgroundAlt.removeAttribute('style')
     }
   }
 }